+++ lib/models/request_data.dart
@@ -5,12 +5,28 @@ import 'package:http_interceptor/extensions/extensions.dart';
 import 'package:http_interceptor/http/http.dart';
 import 'package:http_interceptor/utils/utils.dart';
 
+/// A class that mimics HTTP Request in order to intercept it's data.
 class RequestData {
+  /// The HTTP method of the request.
+  ///
+  /// Most commonly "GET" or "POST", less commonly "HEAD", "PUT", or "DELETE".
+  /// Non-standard method names are also supported.
   Method method;
+
+  /// The base URL String to which the request will be sent. It does not include
+  /// the query parameters.
   String baseUrl;
+
+  /// Map of String to String that represents the headers of the request.
   Map<String, String> headers;
+
+  /// Map of String to String that represents the query parameters of the
+  /// request.
   Map<String, dynamic> params;
-  dynamic? body;
+
+  dynamic body;
+
+  /// The encoding used for the request.
   Encoding? encoding;
 
   RequestData({
@@ -23,24 +39,38 @@ class RequestData {
   })  : headers = headers ?? {},
         params = params ?? {};
 
+  /// The complete URL String including query parameters to which the request
+  /// will be sent.
   String get url => buildUrlString(baseUrl, params);
 
-  factory RequestData.fromHttpRequest(Request request) {
+  /// Creates a new request data from an HTTP request.
+  ///
+  /// For now it only supports [Request].
+  /// TODO(codingalecr): Support for [MultipartRequest] and [StreamedRequest].
+  factory RequestData.fromHttpRequest(BaseRequest request) {
     var params = Map<String, dynamic>();
     request.url.queryParametersAll.forEach((key, value) {
       params[key] = value;
     });
     String baseUrl = request.url.origin + request.url.path;
-    return RequestData(
-      method: methodFromString(request.method),
-      encoding: request.encoding,
-      body: request.body,
-      baseUrl: baseUrl,
-      headers: request.headers,
-      params: params,
+
+    if (request is Request) {
+      return RequestData(
+        method: methodFromString(request.method),
+        baseUrl: baseUrl,
+        headers: request.headers,
+        body: request.body,
+        encoding: request.encoding,
+        params: params,
+      );
+    }
+
+    throw UnsupportedError(
+      "Can't intercept ${request.runtimeType}. Request type not supported yet.",
     );
   }
 
+  /// Converts this request data to an HTTP request.
   Request toHttpRequest() {
     var reqUrl = buildUrlString(baseUrl, params);
 
@@ -50,7 +80,7 @@ class RequestData {
     if (encoding != null) request.encoding = encoding!;
     if (body != null) {
       if (body is String) {
-        request.body = body;
+        request.body = body as String;
       } else if (body is List) {
         request.bodyBytes = body?.cast<int>();
       } else if (body is Map) {
@@ -63,6 +93,7 @@ class RequestData {
     return request;
   }
 
+  /// Convenient toString implementation for logging.
   @override
   String toString() {
     return 'Request Data { $method, $baseUrl, $headers, $params, $body }';